# backend/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "API REST para gestão inteligente de fotografias e eventos caninos do PS (Photo Storage).",
        "title": "PS API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "Suporte PS",
            "email": "suporte@ps.com"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/",
    "paths": {
...
        "/api/v1/reports/clients-csv": {
            "post": {
                "description": "Inicia a extração assíncrona do relatório consolidado de clientes, cães, fotos e pagamentos do tenant autenticado e envia o link para o e-mail cadastrado",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Reports"
                ],
                "summary": "Exportar relatório CSV de clientes",
                "parameters": [
                    {
                        "type": "string",
                        "description": "ID do evento para filtrar o relatório",
                        "name": "season_id",
                        "in": "query"
                    }
                ],
                "responses": {
                    "202": {
                        "description": "Accepted",
                        "schema": {
                            "$ref": "#/definitions/httpx.SuccessEnvelope"
                        }
                    },
                    "401": {
                        "description": "Não autenticado",
                        "schema": {
                            "$ref": "#/definitions/httpx.ErrorEnvelope"
                        }
                    },
                    "403": {
                        "description": "Tenant não associado ou bloqueado",
                        "schema": {
                            "$ref": "#/definitions/httpx.ErrorEnvelope"
                        }
                    },
                    "500": {
                        "description": "Erro interno",
                        "schema": {
                            "$ref": "#/definitions/httpx.ErrorEnvelope"
                        }
                    }
                }
            }
        },
        "/api/v1/reports/clients-pdf": {
            "get": {
                "description": "Gera e faz o download direto instantâneo do relatório consolidado em formato PDF",
                "produces": [
                    "application/pdf"
                ],
                "tags": [
                    "Reports"
                ],
                "summary": "Download direto do relatório PDF de clientes",
                "parameters": [
                    {
                        "type": "string",
                        "description": "ID do evento para filtrar o relatório",
                        "name": "season_id",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Arquivo PDF",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "401": {
                        "description": "Não autenticado",
                        "schema": {
                            "$ref": "#/definitions/httpx.ErrorEnvelope"
                        }
                    },
                    "403": {
                        "description": "Tenant não associado ou bloqueado",
                        "schema": {
                            "$ref": "#/definitions/httpx.ErrorEnvelope"
                        }
                    },
                    "500": {
                        "description": "Erro interno",
                        "schema": {
                            "$ref": "#/definitions/httpx.ErrorEnvelope"
                        }
                    }
                }
            },
            "post": {
                "description": "Inicia a extração assíncrona do relatório consolidado em PDF de clientes, cães e fotos para o tenant autenticado e envia o link para o e-mail cadastrado",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Reports"
                ],
                "summary": "Exportar relatório PDF de clientes",
                "parameters": [
                    {
                        "type": "string",
                        "description": "ID do evento para filtrar o relatório",
                        "name": "season_id",
                        "in": "query"
                    }
                ],
                "responses": {
                    "202": {
                        "description": "Accepted",
                        "schema": {
                            "$ref": "#/definitions/httpx.SuccessEnvelope"
                        }
                    },
                    "401": {
                        "description": "Não autenticado",
                        "schema": {
                            "$ref": "#/definitions/httpx.ErrorEnvelope"
                        }
                    },
                    "403": {
                        "description": "Tenant não associado ou bloqueado",
                        "schema": {
                            "$ref": "#/definitions/httpx.ErrorEnvelope"
                        }
                    },
                    "500": {
                        "description": "Erro interno",
                        "schema": {
                            "$ref": "#/definitions/httpx.ErrorEnvelope"
                        }
                    }
                }
            }
        },
        "/api/v1/reports/download": {
            "get": {
                "description": "Permite o download seguro de um arquivo de relatório previamente gerado, validando isolamento de tenant",
                "produces": [
                    "text/csv",
                    "application/pdf"
                ],
                "tags": [
                    "Reports"
                ],
                "summary": "Baixar arquivo de relatório gerado",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Caminho relativo do arquivo de relatório",
                        "name": "file",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Arquivo de relatório",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "Parâmetro inválido",
                        "schema": {
                            "$ref": "#/definitions/httpx.ErrorEnvelope"
                        }
                    },
                    "401": {
                        "description": "Não autenticado",
                        "schema": {
                            "$ref": "#/definitions/httpx.ErrorEnvelope"
                        }
                    },
                    "403": {
                        "description": "Acesso não autorizado ao arquivo",
                        "schema": {
                            "$ref": "#/definitions/httpx.ErrorEnvelope"
                        }
                    },
                    "404": {
                        "description": "Arquivo não encontrado",
                        "schema": {
                            "$ref": "#/definitions/httpx.ErrorEnvelope"
                        }
                    },
                    "500": {
                        "description": "Erro interno",
                        "schema": {
                            "$ref": "#/definitions/httpx.ErrorEnvelope"
                        }
                    }
                }
            }
        },
        "/api/v1/reports/paid-clients-csv": {
            "post": {
                "description": "Inicia a extração assíncrona do relatório de clientes e fotos com pagamentos confirmados (pagos) do tenant autenticado e envia o link para o e-mail cadastrado",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Reports"
                ],
                "summary": "Exportar relatório CSV de clientes com pagamentos confirmados",
                "parameters": [
                    {
                        "type": "string",
                        "description": "ID do evento para filtrar o relatório",
                        "name": "season_id",
                        "in": "query"
                    }
                ],
                "responses": {
                    "202": {
                        "description": "Accepted",
                        "schema": {
                            "$ref": "#/definitions/httpx.SuccessEnvelope"
                        }
                    },
                    "401": {
                        "description": "Não autenticado",
                        "schema": {
                            "$ref": "#/definitions/httpx.ErrorEnvelope"
                        }
                    },
                    "403": {
                        "description": "Tenant não associado ou bloqueado",
                        "schema": {
                            "$ref": "#/definitions/httpx.ErrorEnvelope"
                        }
                    },
                    "500": {
                        "description": "Erro interno",
                        "schema": {
                            "$ref": "#/definitions/httpx.ErrorEnvelope"
                        }
                    }
                }
            }
        },
        "/api/v1/reports/unpaid-clients-csv": {
            "post": {
                "description": "Inicia a extração assíncrona do relatório de clientes e fotos com pagamentos não quitados ou pendentes do tenant autenticado e envia o link para o e-mail cadastrado",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Reports"
                ],
                "summary": "Exportar relatório CSV de clientes com pagamentos não pagos",
                "parameters": [
                    {
                        "type": "string",
                        "description": "ID do evento para filtrar o relatório",
                        "name": "season_id",
                        "in": "query"
                    }
                ],
                "responses": {
                    "202": {
                        "description": "Accepted",
                        "schema": {
                            "$ref": "#/definitions/httpx.SuccessEnvelope"
                        }
                    },
                    "401": {
                        "description": "Não autenticado",
                        "schema": {
                            "$ref": "#/definitions/httpx.ErrorEnvelope"
                        }
                    },
                    "403": {
                        "description": "Tenant não associado ou bloqueado",
                        "schema": {
                            "$ref": "#/definitions/httpx.ErrorEnvelope"
                        }
                    },
                    "500": {
                        "description": "Erro interno",
                        "schema": {
                            "$ref": "#/definitions/httpx.ErrorEnvelope"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "httpx.ErrorEnvelope": {
            "type": "object",
            "properties": {
                "errors": {},
                "message": {
                    "type": "string"
                },
                "success": {
                    "type": "boolean"
                }
            }
        },
        "httpx.SuccessEnvelope": {
            "type": "object",
            "properties": {
                "data": {},
                "success": {
                    "type": "boolean"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "Digite \"Bearer \" seguido do token JWT obtido no login.",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}